requirements.txt
Flask==2.3.2
google-cloud-vision==3.5.2
vertexai==1.95.1
gunicorn==20.1.0
PyYAML
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
docstring_parser==0.17.0
executing==2.2.1
fastjsonschema==2.21.2
gcloud==0.18.3
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.129.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-genai==1.53.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.3.5
oauth2client==4.1.3
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
shapely==2.1.2
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
tenacity==9.1.2
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.0
visionai @ file:///home/kanak/Documents/josie/visionai-0.0.5-py3-none-any.whl#sha256=b862b7edcbd2a2b5d8385dfebe7740834c55c6b3d7b9d5b88276f9b402777a26
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
yarg==0.1.9
